refactor: Classes organized into packages to improve clarity and facilitate maintenance
